requirements.txt
mysql-connector-python
nltk


